# docs/openapi.json
{
  "components": {
    "schemas": {
      "ArtifactBundle": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "descriptor": {
            "$ref": "#/components/schemas/ModelDescriptor"
          },
          "digest": {
            "title": "Digest",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ArtifactFile"
            },
            "title": "Files",
            "type": "array"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "validation_state": {
            "default": "unverified",
            "enum": [
              "validated",
              "failed",
              "unverified"
            ],
            "title": "Validation State",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "digest",
          "files",
          "descriptor",
          "manifest"
        ],
        "title": "ArtifactBundle",
        "type": "object"
      },
      "ArtifactFile": {
        "additionalProperties": false,
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "path",
          "size_bytes",
          "sha256"
        ],
        "title": "ArtifactFile",
        "type": "object"
      },
...
      "CapabilitySnapshot": {
        "additionalProperties": false,
        "properties": {
          "adapters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Adapters",
            "type": "array"
          },
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "environment": {
            "additionalProperties": true,
            "title": "Environment",
            "type": "object"
          },
          "fingerprint": {
            "title": "Fingerprint",
            "type": "string"
          },
          "packages": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Packages",
            "type": "object"
          },
          "probes": {
            "additionalProperties": true,
            "title": "Probes",
            "type": "object"
          },
          "providers": {
            "items": {
              "type": "string"
            },
            "title": "Providers",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "fingerprint",
          "checked_at",
          "packages",
          "environment",
          "providers",
          "probes",
          "adapters"
        ],
        "title": "CapabilitySnapshot",
        "type": "object"
      },
...
      "ModelDescriptor": {
        "additionalProperties": false,
        "properties": {
          "blockers": {
            "items": {
              "type": "string"
            },
            "title": "Blockers",
            "type": "array"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/TensorSpec"
            },
            "title": "Inputs",
            "type": "array"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/TensorSpec"
            },
            "title": "Outputs",
            "type": "array"
          },
          "runtime_compatibility": {
            "enum": [
              "compatible",
              "blocked",
              "unknown_until_probe"
            ],
            "title": "Runtime Compatibility",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/ModelState"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "name",
          "state",
          "runtime_compatibility"
        ],
        "title": "ModelDescriptor",
        "type": "object"
      },
      "ModelState": {
        "additionalProperties": false,
        "properties": {
          "architecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "architecture_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture Version"
          },
          "domains": {
            "items": {
              "type": "string"
            },
            "title": "Domains",
            "type": "array"
          },
          "format": {
            "enum": [
              "pytorch-bundle",
              "onnx",
              "safetensors"
            ],
            "title": "Format",
            "type": "string"
          },
          "opsets": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Opsets",
            "type": "object"
          },
          "optimized": {
            "default": false,
            "title": "Optimized",
            "type": "boolean"
          },
          "precision": {
            "default": "unknown",
            "enum": [
              "fp32",
              "dynamic-int8",
              "static-int8",
              "fp16",
              "unknown"
            ],
            "title": "Precision",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "runtime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime"
          }
        },
        "required": [
          "format"
        ],
        "title": "ModelState",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "request_id",
          "retryable",
          "details"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "TensorSpec": {
        "additionalProperties": false,
        "properties": {
          "bounds": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "title": "Bounds",
            "type": "object"
          },
          "class_axis": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Axis"
          },
          "dimensions": {
            "items": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "maxItems": 8,
            "title": "Dimensions",
            "type": "array"
          },
          "dtype": {
            "enum": [
              "float32",
              "float64",
              "float16",
              "int64",
              "int32",
              "int8",
              "uint8",
              "bool"
            ],
            "title": "Dtype",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z_][A-Za-z0-9_.-]*$",
            "title": "Name",
            "type": "string"
          },
          "shape": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shape"
          },
          "value_range": {
            "anyOf": [
              {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Range"
          }
        },
        "required": [
          "name",
          "dtype",
          "dimensions"
        ],
        "title": "TensorSpec",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Local single-operator model workflows. Bearer authentication required. Expensive operations return durable jobs; follow /jobs/{id}/events with Authorization and Last-Event-ID headers.",
    "title": "ModelPort",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitySnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Capabilities"
      }
    },
...
    "/api/v1/models": {
      "get": {
        "operationId": "models_api_v1_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ArtifactBundle"
                  },
                  "title": "Response Models Api V1 Models Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Models"
      }
    },
...
  }
}